* new proxy structure
* protecting ENS registration + adding initialization without ENS (for
upgrade of preregistered proxy)

adding erc1271 capability to wallet + debug

mechanism as well ?

removing upgradability features as requested by marek

fix linter
