--- universal-login-contracts/contracts/WalletMasterCopy.sol
@@ -1,11 +1,0 @@
-pragma solidity ^0.5.2;
-
-import "./common/MasterCopy.sol";
-import "./WalletContract.sol";
-
-
-contract WalletMasterCopy is MasterCopy, WalletContract {
-    constructor(address _key, bytes32 _hashLabel, string memory _name, bytes32 _node, ENS ens, FIFSRegistrar registrar, PublicResolver resolver)
-        WalletContract(_key, _hashLabel, _name, _node, ens, registrar, resolver) public {
-    }
-}
