--- universal-login-contracts/contracts/common/SelfAuthorized.sol
@@ -1,12 +1,0 @@
-//TODO: Before release update to last audited version
-pragma solidity ^0.5.0;
-
-
-/// @title SelfAuthorized - authorizes current contract to perform actions
-/// @author Richard Meissner - <richard@gnosis.pm>
-contract SelfAuthorized {
-    modifier authorized() {
-        require(msg.sender == address(this), "Method can only be called from this contract");
-        _;
-    }
-}
