+++ universal-login-contracts/contracts/erc1836/interfaces/IERC1271.sol
@@ -1,0 +1,6 @@
+pragma solidity ^0.5.0;
+
+
+contract IERC1271 {
+    function isValidSignature(bytes32, bytes memory) public view returns (bool isValid);
+}
