+++ universal-login-contracts/contracts/mock/MockWalletMaster.sol
@@ -1,9 +1,9 @@
 pragma solidity ^0.5.0;
 
-import "../common/MasterCopy.sol";
+import "../WalletMaster.sol";
 
 
-contract MockMasterCopy is MasterCopy {
+contract MockWalletMaster is WalletMaster {
     uint256 _count;
 
     function giveAway() external payable {
