+++ universal-login-contracts/package.json
@@ -26,7 +26,7 @@
     "ganache-core": "^2.2.1",
     "mocha": "^5.2.0",
     "openzeppelin-solidity": "^2.1.1",
-    "solium": "^1.1.8",
+    "ethlint": "^1.2.3",
     "ts-node": "^8.0.2",
     "tslint": "^5.12.1",
     "typescript": "^3.3.3"
