+++ universal-login-contracts/test/contracts/ENSRegistered.js
@@ -1,10 +1,12 @@
 import chai, {expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import ENSRegistered from '../../build/ENSRegistered';
+import WalletMaster from '../../build/WalletMaster';
+import Proxy from '../../build/Proxy';
 import ENSBuilder from 'ens-builder';
 import {createMockProvider, deployContract, getWallets, solidity} from 'ethereum-waffle';
 import {utils} from 'ethers';
 import {lookupAddress, withENS} from '../utils';
+
 
 chai.use(chaiAsPromised);
 chai.use(solidity);
@@ -15,11 +17,12 @@ const hashLabel = utils.keccak256(utils.toUtf8Bytes(label));
 const name = `${label}.${domain}`;
 const node = utils.namehash(name);
 
-describe('WalletContract contract', async () => {
+describe('WalletContract', async () => {
   let provider;
   let wallet;
   let ensBuilder;
-  let walletContract;
+  let walletMaster;
+  let walletProxy;
 
   beforeEach(async () => {
     provider = createMockProvider();
@@ -29,11 +32,14 @@ describe('WalletContract contract', async () => {
     provider = withENS(provider, ensAddress);
     const registrar = ensBuilder.registrars[domain].address;
     const args = [hashLabel, name, node, ensBuilder.ens.address, registrar, ensBuilder.resolver.address];
-    walletContract = await deployContract(wallet, ENSRegistered, args);
+    walletMaster = await deployContract(wallet, WalletMaster, []);
+    const initData = new utils.Interface(WalletMaster.interface).functions.initializeWithENS.encode([wallet.address, ...args]);
+    walletProxy = await deployContract(wallet, Proxy, [walletMaster.address, initData]);
   });
 
+
   it('resolves to given address', async () => {
-    expect(await provider.resolveName('alex.mylogin.eth')).to.eq(walletContract.address);
-    expect(await lookupAddress(provider, walletContract.address, ensBuilder.resolver.address)).to.eq('alex.mylogin.eth');
+    expect(await provider.resolveName('alex.mylogin.eth')).to.eq(walletProxy.address);
+    expect(await lookupAddress(provider, walletProxy.address, ensBuilder.resolver.address)).to.eq('alex.mylogin.eth');
   });
 });
