+++ universal-login-contracts/test/fixtures/basicMasterAndProxy.js
@@ -1,15 +1,17 @@
 import {deployContract} from 'ethereum-waffle';
 import MockToken from '../../build/MockToken';
-import MockMasterCopy from '../../build/MockMasterCopy';
-import MockProxy from '../../build/MockProxy';
+import MockWalletMaster from '../../build/MockWalletMaster';
+import Proxy from '../../build/Proxy';
+import {Contract} from 'ethers';
 
 export default async function basicMasterAndProxy(provider, [, , , , , , , , , wallet]) {
   const publicKey = wallet.address;
   const keyAsAddress = wallet.address;
   const {provider} = wallet;
   const privateKey = wallet.privateKey;
-  const masterCopy = await deployContract(wallet, MockMasterCopy);
-  const proxy = await deployContract(wallet, MockProxy, [masterCopy.address]);
+  const identityMaster = await deployContract(wallet, MockWalletMaster);
+  const identityProxy = await deployContract(wallet, Proxy, [identityMaster.address, []]);
   const mockToken = await deployContract(wallet, MockToken);
-  return {provider, publicKey, privateKey, keyAsAddress, masterCopy, proxy, mockToken, wallet};
+  const proxyAsIdentity = new Contract(identityProxy.address, MockWalletMaster.abi, wallet);
+  return {provider, publicKey, privateKey, keyAsAddress, identityMaster, identityProxy, proxyAsIdentity, mockToken, wallet};
 }
