+++ universal-login-relayer/test/relayer/routes/authorisations.js
@@ -3,7 +3,7 @@ import chaiHttp from 'chai-http';
 import {RelayerUnderTest} from '../../../lib/utils/relayerUnderTest';
 import {createMockProvider, getWallets} from 'ethereum-waffle';
 import {waitForContractDeploy} from 'universal-login-commons';
-import WalletContract from 'universal-login-contracts/build/WalletContract';
+import WalletContract from 'universal-login-contracts/build/Wallet';
 
 chai.use(chaiHttp);
